I get a 403 in response when I use "slurp", which has a User-Agent of "Java/1.6.0_65"; but I get a 200 response when I use "clj-http.client/get", which has a User-Agent of "Apache-HttpClient/4.3.5 (java 1.5)".  I'm not sure why this is the case, because HN doesn't seem to have a problem with me setting the User-Agent to either "Java" or "foo" as the value.
Amendment: I did a quick search, and learned that sites have been specifically
blocking User-Agents of the form "Java/xyz" for a while now, because this is
typically the signature of bots that don't want to play nice and identify
themselves.  In light of this, I find it strange that Clojure doesn't set the
user-agent to something else for its "slurp" function.